pyproject.toml
[project]
name = "day-trader"
version = "0.1.0"
description = "Evolutionary algorithmic trading arena"
requires-python = ">=3.12"
dependencies = [
    "alpaca-py>=0.33.1",
    "anthropic>=0.39",
    "sqlalchemy>=2.0",
    "pandas>=2.2",
    "pandas-ta>=0.3.14b",
    "numpy>=1.26",
    "scikit-learn>=1.5",
    "deap>=1.4",
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "websockets>=13",
    "apscheduler>=3.10,<4",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "pytz>=2024.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"
